[master]采用一种新方式去处理点9图片，基本原理是记录变化区域，然后用系统方法一次次去拉伸，优点是解决了以前图片拼装的时候图片不完整的问题； 另外图片平装定方式也做了优化，采用固定区域前后多绘制一帧的方式，解决了多点变化图片有缝隙的问题

+++ NinePointsImageDemo/NinePointsImageDemo/Models/PointLocation.h
@@ -11,5 +11,5 @@
 
 @property (nonatomic, assign) float startLocation;
 @property (nonatomic, assign) float endLocation;
-@property (nonatomic, assign) float length;
+@property (nonatomic, assign) int length;
 @end
